requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
psycopg2-binary==2.9.6